requirements.txt
opencv-python==4.3.0.38
torch==2.3.1
torchvision==0.18.1
torchsummary==1.5.1
pytorch-lightning==2.3.1
PyYAML <=6.0.1
loguru==0.7.2
tqdm==4.66.4
tensorboard==2.14.0
protobuf==4.25.3
scikit-image==0.21.0
pandas==2.0.3
matplotlib==3.7.5
scikit-learn==1.3.2
albumentations==1.4.11
huggingface_hub==0.24.0
onnx==1.16.2
onnxruntime-gpu==1.18.1
